make project structure better
