+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "asymmetric-crypto"
 version = "0.1.0"
-authors = ["tiannian <dtiannian@aliyun.com>"]
+authors = ["tiannian <dtiannian@aliyun.com>", "xujian009 <s1473561555@sina.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -9,8 +9,10 @@ edition = "2018"
 [dependencies]
 hex = { version = "0.4", default-features = false }
 rand = "0.7.3"
-tiny-keccak = { version = "2.0.0", features = ["sha3"] }
 dislog-hal = { git = "https://github.com/Yinet-project/dislog-hal"}
+
+[dev-dependencies]
+tiny-keccak = { version = "2.0.0", features = ["sha3"] }
 dislog-hal-curve25519 = { git = "https://github.com/Yinet-project/dislog-hal-curve25519"}
 dislog-hal-sm2 = { git = "https://github.com/Yinet-project/dislog-hal-sm2"}
 curve25519-dalek = "2"
